+++ src/app/routes/router.tsx
@@ -5,8 +5,8 @@ import { createBrowserRouter, Navigate } from "react-router";
 import { OauthCallBack } from "@/pages/auth/ui/oauth-callback";
 import { CheckupResultPage } from "@/pages/checkup-result/ui/checkup-result";
 import { HealthReportDetailPage } from "@/pages/health-report/health-report-detail";
-import { HealthTipPage } from "@/pages/health-tip/health-tip";
-import { HealthTipDetailPage } from "@/pages/health-tip/health-tip-detail";
+import { HealthTipPage } from "@/pages/health-tip/ui/health-tip";
+import { HealthTipDetailPage } from "@/pages/health-tip/ui/health-tip-detail";
 import { HomePage } from "@/pages/home/ui/home";
 import { LoginPage } from "@/pages/login/ui/login";
 import { MyPage } from "@/pages/my-page/my-page";
