+++ src/pages/health-tip/ui/health-tip-detail.tsx
